[feat] deleteTodos 선택 중복 삭제 구현

+++ src/components/Form.jsx
@@ -1,26 +1,51 @@
 import List from "./List";
-import { useState } from "react";
+import { useState, useEffect } from "react";
 import { useDispatch } from "react-redux";
 import styled from "styled-components";
 import { createTodos } from "../redux/modules/todos";
 import { deleteTodos } from "../redux/modules/todos";
 const Form = () => {
+  const dispatch = useDispatch();
   const [text, setText] = useState("");
+  const [date, setDate] = useState("");
   const [deadLine, setDeadLine] = useState("");
   const [checkedItems, setCheckedItems] = useState([]);
-  const dispatch = useDispatch();
-  let req = {
-    text,
-    deadLine,
-  };
+
+  // 오늘 날짜 밀리세컨으로 변환
+  const timeStamp = new Date();
+  const year = timeStamp.getFullYear();
+  const month = ("0" + (1 + timeStamp.getMonth())).slice(-2);
+  const day = ("0" + timeStamp.getDate()).slice(-2);
+  const today = year + "-" + month + "-" + day;
+  const todayMil = Date.parse(today);
+  // 선택된 날짜 밀리세컨으로 변환
+  const selectedDate = Date.parse(deadLine);
+
+  useEffect(() => {
+    if (selectedDate < todayMil) {
+      alert("현재 보다 이전의 날짜는 설정할 수 없습니다.");
+      setDeadLine("");
+    }
+  }, [deadLine]);
 
   const onSubmitHandler = async (e) => {
+    let req = {
+      text,
+      deadLine,
+    };
     e.preventDefault();
-    await Promise.all(dispatch(createTodos(req)));
+    dispatch(createTodos(req));
   };
+  console.log(checkedItems);
   const onDeleteHandler = () => {
     if (checkedItems.length > 0) {
-      dispatch(deleteTodos(checkedItems));
+      if (window.confirm("삭제할까요?") == true) {
+        dispatch(deleteTodos(checkedItems));
+        setCheckedItems([]);
+      } else {
+        setCheckedItems([]);
+        return false;
+      }
     }
   };
   return (
